Install and save packages given on command line
Now it is possible to call it like that:

    npm-git install --save \
      git@github.com:sample-package-1.git \
      https://github.com/sample-package-2.git#tag-or-branch-or-sha

This will install the packages, determine sha of last commit that matches given revision (master by default) and save this to package.json gitDependencies section.

Next time one calls

    npm git install

the same revisions of the packages will be installed.

This way one can have predictable deplpyments, even if new commits are added to the repo after initial installation. This is an alternative to #10.